Fix type error between localstorage and elm
Yeah, of course there's a bug here. We sent an array, and expected
to only get back a single json encoded value, but in fact we're getting
all of them back.

This way we just tell JS which value to return back to us, so it's
quite a bit easier.