Default some text to 16px
On mobile safari, if an input has a font size of
less than 16px, the browser will zoom when the control
receives focus, which is pretty jarring most of the time.

Rather than deal with that, making the webpage feel more
native and respecting a reasonable font size is a much better
solution, in my humble opinion.

For more information, see https://stackoverflow.com/questions/2989263/disable-auto-zoom-in-input-text-tag-safari-on-iphone