Version the appcache, not the assets
Yeah, so I'm just going to link this here and maybe express some
regrets. Ragrets.

https://alistapart.com/article/application-cache-is-a-douchebag

Basically, If your cache includes an asset, then the browser is going
to assume that asset will continue to exist forever. When we versioned
assets, the next time a deploy occurred, a browser that had previously
visited the site would attempt to fetch the resources for the old
deploy, which naturally would not exist on our instance. Now we keep the
file names the same and allow the browser to decide when and if it
should update.

Signed-off-by: Tim Jarratt <tjarratt@pivotal.io>