Extract AddPhrase form into its own component
Also refactored the "currentActivity" field to be more clear. Where we
had once captured the "activity" that a user was doing, we have grown to
have a need for a "currentPgae". Some pages have activities that are
used to configure components so they know which labels to display, which
URLS to make requests to, etc, etc, etc...

Signed-off-by: Tim Jarratt <tjarratt@pivotal.io>