Extract an AddPhrase component out of the app module
App modules was getting too big. This was clear and evident from the experiences
I was having trying to continue growing it. Additionally, I found that there were
several impossible states in the application's model that were possible, but
didn't actually make sense.

For example, I think it was possible to be adding a phrase, but for the user's
uuid to still be Nothing. In practice, this didn't happen because we ask for
the user's uuid at the beginning, before anything else happens. There should
be a lot less useless checking of Maybes in our updates now, because we are
only keeping references to things that require certain data (like a user uuid
or an activity) when those are actually available, which is pretty rad.

This was a much, much, much bigger refactor than I intended to do. Initially I
thought that it would be possible to simply extract the salient parts parts of the
"practice phrases" model and update functions into another module and immediately
start reaping the benefits.

In practice ...

* There were type changes
  -> I coupled the refactor to removals of Maybes
* There were architecture changes necessitated by my test framework
  -> Turns out that Elmer cannot currently call fakes in modules other than
     the one your subject (app/component) is in. That necessitated a LOT of
     debugging to explain why the focus task fakes stopped working.
* There were structural questions to answer about components
  -> Evan and Elm doesn't have a proposed solution for SPA or components because
     that's not a design goal of Elm. Here, it seemed necessary so I had to do
     some of the necessary research to see how to chain updates and nest models
     and nest messages, but in practice all of that made the extraction a lot
     messier.
* The tests were not sufficient to cover a full refactor
  -> some of the tests were not testing with a realistic setup.
     because they were written one at a time, as the app grew organically,
     and there were a lot of features and application state added over time,
     not all of the tests respected various things like the user actually being
     logged in. Doing something as simple as forcing the user to be logged in
     before showing a component necessitated rewriting a lot of tests.

In practice, I'd like to be able to do this simpler, in the style of
99 Bottles of OO, by Sandi Metz. Next time I should try backing out of the
refactor sooner, and backfilling tests, or looking for smaller changes that
can be done incrementally without breaking the tests or changing the architecture
in a really massive way.