Add nginx config to enable SPA routing
Yeah actually I don't understand this and that's kind of startling and
maybe even a bit frightening ?

This is in effort to enable SPA routing on CF. What I want is for the SPA
routing that I've created to work as intended, and then when the user refreshes
the page, they should be able to fetch the exact same non-root URL and continue
using the application from that point.

Now, if you're reading this, you're probably wondering "well, Tim, it sure seems
like you wrote a pretty fancy nginx config" that did this. And, well, you'd be
wrong.

You'd be wrong because I got this nginx config by downloading the files from the
most recent push of my application and copying out the nginx.conf.

Then I copied it into the root of the app during the deploy, and hey, it just works.

It shouldn't. It should actually behave in the same way it did before. Can't say
that I understand the config well enough to understand why. Probably some feature
getting disabled ?